fix: 修复modules 包没有安装fs-extra依赖的问题
